+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `padding` parameter to Panel
 - Added 'indeterminate' state when progress bars aren't started
+- Fixed Progress deadlock https://github.com/willmcgugan/rich/issues/90
 
 ## [1.2.2] - 2020-05-22